refactor: "워드클라우드 부분 리팩토링 및 기능 추가"
