+++ src/pages/WordCloudPage.js
@@ -1,23 +1,15 @@
 import React from "react";
 import {useNavigate} from "react-router-dom";
 import PropTypes from "prop-types";
-import WordCloudSection from "../components/wordcloud-components/WordCloudSection";
-import StyledWrapper from "../style/StyledWordCloudPage";
+import WordCloudSection from "../components/WordCloud/WordCloudSection";
+import {StyledWrapper} from "../style/StyledWordCloudPage";
 
 const WordCloudPage = () => {
   const navigate = useNavigate();
 
-  const setSearch = (data) => {
-    navigate("/bills", {
-      state: {
-        inputValue: data.name,
-        age: data.age,
-      },
-    });
-  };
   return (
     <StyledWrapper>
-      <WordCloudSection setSearch={setSearch} size={100} />
+      <WordCloudSection size={100} />
       <button type="button" className="compare-button" onClick={() => navigate("/compare")}>
         비교해보기
       </button>
