+++ src/style/StyledComparedWordCloudPage.js
@@ -24,5 +24,3 @@ export const StyledWrapper = styled.div`
     }
   }
 `;
-
-export default StyledWrapper;
