+++ src/style/StyledWordCloud.js
@@ -14,5 +14,3 @@ export const StyledWrapper = styled.div`
     }
   }
 `;
-
-export default StyledWrapper;
