+++ src/style/StyledWordCloudPage.js
@@ -31,5 +31,3 @@ export const StyledWrapper = styled.div`
     }
   }
 `;
-
-export default StyledWrapper;
