--- src/wordcloud_data/billsNouns/16th-2000-nouns.txt
@@ -1,1 +1,0 @@
-1980년, 가족생활, 감사, 감사, 감사, 감정, 감축, 강제, 강제, 강제, 개량, 개발, 개선, 개선, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개혁, 거래, 거래, 거래, 거주, 거창, 건강, 건강증진, 건설, 건설, 건설, 건전, 건축법, 검찰청, 검찰청, 격차, 경감, 경감, 경기, 경영, 경영, 경제, 경찰, 계획법, 계획법, 고엽제, 고엽제, 고용, 고용, 고용보험법, 고용보험법, 공단, 공무원, 공무원, 공무원법, 공무원법, 공무원법, 공무원법, 공사, 공적자금, 공적자금, 공정, 공중위생, 공직, 공직, 공직자윤리법, 공직자윤리법, 공학, 공학, 공항, 과, 과학기술, 관, 관, 관, 관, 관, 관, 관, 관, 관, 관, 관, 관, 관, 관, 관, 관, 관, 관, 관, 관련, 관련, 관련, 관련자, 관리, 관리, 관리, 관리, 관리, 관리, 관리, 관리, 관리, 관리, 관리법, 관리법, 관리법, 관리법, 관리법, 관리법, 관리법, 관의, 관치, 광고물, 광고물, 광역, 광역, 교류, 교류, 교류, 교원, 교육, 교육, 교육, 교육기본법, 교통, 교통, 교통, 구매, 구제, 구조, 구현, 국가, 국가, 국가, 국가보안법, 국가인권위원회법, 국가채무, 국경일, 국민, 국민건강보험, 국민건강보험, 국민연금, 국민연금, 국유재산, 국유재산, 국자, 국정, 국정, 국정, 국제공항, 국토이용관리법, 국회, 국회법, 국회법, 국회법, 국회법, 국회법, 국회법, 국회법, 국회법, 국회법, 국회법, 국회법, 국회법, 국회의원, 군인, 군인, 군인, 권법, 규명, 규제, 규제, 근로기준법, 근로기준법, 근로기준법, 근로기준법, 금, 금융, 금융, 금융, 금지, 금지, 급, 급여, 기관, 기금, 기금, 기금, 기금, 기념사업회, 기능, 기본, 기본, 기본, 기본법, 기본법, 기본법, 기본법, 기본법, 기본법, 기본법, 기술, 기술, 기재, 낙동강, 난치, 남녀, 남녀, 남녀, 남북, 남북, 남북, 남북협력기금, 납북자, 농, 농, 농, 농어촌, 농어촌, 농업, 농촌, 대기, 대기, 대도시권, 대도시권, 대책, 대회, 대회, 댐, 도로, 도로교통법, 도법, 도법, 도법, 독도, 독립, 독립, 독점, 동원, 동학혁명, 등, 등, 등, 등, 등, 등, 등기, 등에, 등에, 등에, 등에, 등에, 등에, 등에, 등에, 등에, 등에, 등에, 등에, 등에, 등에, 등에, 등에, 등에, 등에, 등에, 등의, 등의, 등의, 등의, 망, 매장, 명예, 명예, 묘지, 문화예술, 문화예술, 문화재, 물관리, 미용사, 민, 민법, 민원사무, 민원사무, 민족, 민주, 민주화운동기념사업회, 발전, 발전, 발전, 방문, 방송법, 방지법, 방지법, 배상, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법령, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법안, 법안, 법안, 법안, 법안, 법안, 법안, 법안, 법안, 법안, 법안, 법안, 법안, 법안, 법안, 법안, 법안, 법안, 법안, 법안, 법안, 법안, 법안, 법안, 법적, 변, 병역법, 병역법, 병역법, 보건, 보건, 보건, 보건, 보건, 보상, 보상금, 보상금, 보수, 보장, 보전, 보전, 보전, 보전, 보호, 보호, 보호, 보호, 보호, 보훈, 복무, 복지, 부당, 부동산, 부산, 부실, 부여, 부재, 부정, 부정, 부정, 부채, 부채, 부패, 부패방지법, 비밀, 사건, 사건, 사건, 사법, 사업, 사업, 사업, 사회복지공동모금회, 사회복지사, 산, 산림법, 산업, 산업, 산업, 산업화, 상법, 상속세, 생명, 생명, 생산, 생활, 서비스, 선거, 선거, 선거, 선거, 선거권, 선고, 설립, 설립, 설치, 성질, 세법, 세법, 소기업, 소도, 소득세법, 소득세법, 소득세법, 소방, 소비, 소송, 소유권, 손해, 송아지, 수, 수계, 수당, 수도권, 수도권, 수산업협동조합, 수질, 시국, 식품, 식품위생법, 실명, 실명, 실효, 아동, 아시아, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안전, 안정, 안정, 안정, 약사법, 약사법, 약사법, 약사법, 양여, 어업, 어업, 어업, 어업, 어업, 어업, 어업, 업, 여, 여객, 여객, 여성, 연금, 염, 염, 영예, 예금자, 예금자, 예비군, 예산, 예산, 예산회계법, 예우, 예우, 예우, 예우, 옥외, 옥외, 온천, 외국인, 외동, 외자, 운수, 운수, 운영, 운영, 운영, 위원회, 유공자, 유공자, 유공자, 유공자, 유니버시아드, 유료, 육성, 육성, 육성, 은행, 읍, 의료, 의료, 의료, 의료, 의료법, 의증, 의증, 이산가족, 이용, 이용, 이용, 이용, 이전, 인사청문회, 인사청문회, 일제, 임시, 임용, 자, 자동차, 자동차, 자동차, 자산, 자의, 장기, 장기, 장수, 장애인, 재정, 재정, 재해, 적자, 전기통신사업법, 전상훈, 전자, 전자, 정리법, 정보, 정보, 정보, 정보, 정보, 정보원, 정보원, 정부, 정부조직법, 정부조직법, 정부조직법, 정비, 정비, 정지원, 정지원, 제정, 제품, 제한, 제한, 제한, 제한, 제한, 제한, 조사, 조사, 조사, 조세, 조세, 조세, 조세, 조세, 조세, 조치, 조치, 조치, 조치, 조치, 조치, 조치, 조치, 조치, 조치, 조치, 조치, 조치, 조치, 조치, 주민등록, 주변, 주택, 주택, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중대장, 중소기업, 중소기업, 중소기업, 증권, 증권거래법, 증권거래법, 증권거래법, 증권거래법, 증언, 증여, 증진, 지급, 지급, 지방, 지방, 지방, 지방, 지방세법, 지방세법, 지방세법, 지방세법, 지방세법, 지방세법, 지방자치법, 지방자치법, 지역, 지역, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지위, 진상, 진흥, 진흥, 진흥, 집단, 징용, 차별, 참전, 참전, 창군, 채용, 책임, 처리, 처리, 처리, 청산, 청소년, 체결, 체결, 촉진, 촉진, 촉진, 촉진, 촉진, 촉진, 촉진, 촉진, 축산, 축소, 출입국, 태평양, 토양, 통신, 통신망, 통신망, 통신망, 통신망, 통신비밀보호법, 통신비밀보호법, 퇴직, 퇴직, 퇴직, 퇴출, 특례, 특례, 특례, 특례, 특례, 특례, 특례법, 특례법, 특례법, 특별, 특별, 특별, 특별, 특별, 특별, 특별, 특별, 특별, 특별, 특별, 특별, 특별, 특별, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별회계, 파산법, 판매, 판매, 평등, 평등, 폐기물관리법, 폐지, 폐지, 포의, 표준시, 피해자, 하, 하계, 하수, 하수, 학교, 학원, 한, 한, 한, 한국, 한국마사회, 함평, 해소, 해직, 헌법재판소, 혁신, 혈액, 협력, 협력, 협정, 협정, 형, 형사소송법, 호적법, 환경, 환경, 환경, 환경, 환경오염, 환자, 환자, 환자, 회, 회, 회계, 회계, 회복, 회복, 회사, 효율, 후유, 후유, 희생자,
