--- src/wordcloud_data/billsNouns/16th-2004-nouns.txt
@@ -1,1 +1,0 @@
-개발, 개인정보, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 개정, 거래, 건립, 경기, 공, 공기, 공무원법, 공무원법, 관, 관, 관의, 교육, 교육, 교육감, 국가, 국가배상법, 국립, 국민건강보험, 국민건강보험, 국민연금, 국회, 국회도서관, 국회법, 국회법, 국회법, 국회법, 금융, 도, 도, 등에, 등에, 민족, 방송법, 방지, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법안, 병역법, 보고, 보호, 보호, 부정, 북, 사면, 사무처, 선거, 선거, 설치, 센터, 신용정보, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 용, 이북, 이용, 인력, 일반, 자치, 자치, 재산, 정보, 정부조직법, 정부조직법, 제한, 제한, 조세, 조세, 조치, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중, 중소기업, 지방, 지방, 지방, 지역, 지원, 지원, 치매, 통신비밀보호법, 특례, 특례, 특별, 특별법, 특별법, 특별법, 특정, 특허법, 폐광, 행위자, 환수,
