--- src/wordcloud_data/billsNouns/17th-2008-nouns.txt
@@ -1,1 +1,0 @@
-가능, 가축, 가축, 감독, 감독, 감사원, 개발, 개발, 개정, 개정, 거래, 거래, 거래법, 거창, 건설, 건축법, 건축법, 검사, 경기, 경기, 경제, 경제, 경제자유구역, 경호실, 계획, 계획, 고도, 고령, 고령, 고용, 고용보험법, 고용보험법, 공공기관의 정보공개에 관한 법률, 공로, 공무원, 공무원법, 공무원법, 공무원법, 공원, 공유, 공유재산, 공익사업, 공인, 공인, 공주, 공중위생, 공직, 공직, 공직자윤리법, 공항, 과학기술, 과학기술, 과학기술, 관, 관, 관, 관련, 관련, 관련, 관련자, 관련자, 관리, 관리, 관리, 관리법, 관리법, 관리법, 관리법, 관리법, 관리법, 관리법, 관리법, 관리법, 광우, 교육, 교육국, 교육법, 구매, 구세계, 구세계, 구제, 구조조정, 국가, 국가, 국가, 국가과학기술자문회의, 국가안전보장회의, 국가인권위원회법, 국립묘지, 국민, 국민, 국민, 국민건강보험, 국민건강보험, 국민연금, 국세, 국적법, 국토, 국토, 국토, 국회법, 군복, 권, 권리, 권익, 규명, 규명, 규제, 균형발전, 근로기준법, 근로자, 금융, 금융, 금융, 금지, 기간제 및 단시간근로자 보호 등에 관한 법률, 기공, 기구, 기구, 기금법, 기반, 기반, 기반, 기반, 기본법, 기본법, 기본법, 기본법, 기본법, 기본법, 기본법, 기본법, 기본법, 기본법, 기본법, 기업, 남녀, 남해안, 노근리, 노인, 놀이, 농, 농, 농림, 농민, 농산물, 농산물, 농어촌, 뇌, 단속, 대, 대, 대구경북, 대비, 대책, 대책, 대통령, 대회, 대회, 도로교통법, 도서관, 도서관, 도시, 도시, 도축장, 독점규제 및 공정거래에 관한 법률, 동, 동학, 등, 등, 등, 등, 등, 등, 등, 등, 등, 등, 등, 등, 등, 등, 등에, 등에, 등에, 론스타, 면상, 명예, 명예, 명예, 명예, 명예, 명예, 무역, 문화, 문화재, 문화재, 물, 물질, 물품, 미술관, 민간, 민방위, 민주화, 민주화, 박람회, 박물관, 발전, 발전, 발전, 발전, 발전, 방송, 범죄, 범죄자, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법안, 법안, 법인, 베이, 병, 보건, 보건, 보건, 보상, 보상, 보상, 보상, 보상, 보상, 보상, 보장, 보조금, 보험업법, 보호, 보호, 보호, 보호, 보호, 복구, 복권, 복권, 복지, 부담금, 부담금, 부착, 북한이탈주민, 분권, 비밀, 비상, 사건, 사건, 사건, 사고, 사고, 사고, 사모펀드, 사업, 사업, 사용자, 사회, 사회, 사회, 사회, 산, 산업, 산업, 산업단지, 삶의 질, 삼성중공업, 삼청교육, 생명, 서, 선거법, 선거법, 선수권대회, 선수권대회, 설립, 설치, 설치, 설치, 설치, 설치, 설치, 성폭력, 성폭력범죄의 처벌 및 피해자보호 등에 관한 법률, 성폭력범죄의 처벌 및 피해자보호 등에 관한 법률, 세계, 세법, 소기업, 소기업, 소득, 소득세법, 소방기본법, 소방시설, 소상, 소상, 소음, 소음, 소해, 손해배상, 손해배상, 쇠고기, 수산업, 수원, 수입, 수행, 순환, 스피리트, 시설, 시설, 시설, 시설, 식품위생법, 신지식인, 신행, 실명, 실업, 심사, 아시아, 아시아, 아시아문화중심도시, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안전, 안전, 안전, 안전, 안전, 안전, 안전, 안정, 어린이, 어업, 어업, 어촌, 여권, 여성, 여성, 여수, 역사, 연구원, 연금, 연기, 영유아보육법, 영유아보육법, 영유아보육법, 예방, 예방법, 예방법, 예산, 예우, 오염, 오염, 외국환, 외환은행, 용산, 운동, 운동, 운영, 운영, 운영, 운영, 운영, 운영, 위원회, 위원회, 위원회, 위치, 위험, 유공자, 유류, 유류, 유류, 유지, 유출, 유출, 유통, 유해, 육상, 육상, 육성, 육성, 윤리, 의혹, 이용, 이용, 인, 인수, 인정, 인천, 인천, 일자리, 임명, 입법, 자, 자동차, 자동차관리법, 자문, 자본시장과 금융투자업에 관한 법률, 자원, 자원, 자의, 자치, 자활, 작전, 장사, 장애인, 장애인, 장치, 재난, 저출산, 저출산, 전기, 전기, 전기용품, 전기통신사업법, 전부, 전부, 전염병, 전염병, 전자, 전자제품, 전쟁, 절차, 절차, 정부조직법, 정부조직법, 정부조직법, 정비, 정수, 정착, 정치자금법, 제작, 제주, 제품, 제한, 제한, 제한, 제한, 제한, 제한, 제화, 조성, 조성, 조성, 조성, 조세, 조세, 조세, 조세, 조치, 조치, 조치, 조치, 조치, 주민, 주민, 주민, 주민, 주민등록, 주택임대차보호법, 중, 중등, 중소기업, 중심, 증, 지방, 지방, 지방, 지방세법, 지방세법, 지방세법, 지속, 지식, 지역, 지역, 지역, 지역, 지역, 지역, 지역, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지적, 지정, 지정, 직업, 진동규, 진상, 진상, 진흥, 진흥, 진흥, 진흥, 진흥, 차별, 차별금지법, 참여자, 참전, 청년, 청소년, 청소년, 체육, 초, 촉진, 촉진, 촉진, 촉진, 최저임금법, 추적, 충돌, 취득, 콘텐츠, 크레인, 토지, 통계, 통신, 투자, 특구, 특구, 특례, 특례, 특례, 특례, 특례법, 특례법, 특별, 특별, 특별, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특수임무수행자, 특정, 특화, 파견, 평등, 폐기물관리법, 폐지, 폐지, 품질, 품질, 피해, 피해, 피해, 피해자, 하도급거래 공정화에 관한 법률, 한국대학교육협의회, 한국전문대학교육협의회, 한국조폐공사, 항공기, 해상, 해상, 해소, 해안, 행정심판법, 행정중심복합도시, 향상, 허, 허가, 혁명, 호, 화성, 화학, 활동, 활동, 활성, 회복, 회복, 회복, 회복, 회복, 회복, 회의, 후방, 후속, 희생자, 희생자,
