--- src/wordcloud_data/billsNouns/18th-2012-nouns.txt
@@ -1,1 +1,0 @@
-가격, 가야, 가정, 가족관계의 등록 등에 관한 법률, 가출, 감염병, 감정평가, 강력범죄, 강제, 개발, 개발, 개발, 개발, 개발, 개별, 개인정보, 개인정보, 거부, 건강, 건물, 건설, 건설, 건설, 검사, 검사, 검사, 검역, 게임, 경륜, 경정, 계획, 고등, 고용, 고용보험법, 골재, 공격, 공무원법, 공사, 공사, 공시, 공인, 공인, 공인, 공장, 공중위생, 공직, 공직, 공직, 공직, 공직, 공직, 공직, 공직, 공항, 과학기술, 과학기술, 관, 관련자, 관리, 관리, 관리, 관리법, 관리법, 관리법, 관리법, 관리법, 관리법, 관리법, 교원, 교원, 교육, 교육법, 교육법, 교육법, 교통안전, 구매, 구제, 국가인권위원회법, 국립대, 국민건강보험, 국민건강증진법, 국민건강증진법, 국민기초생활 보장법, 국유재산, 국토, 국회의원, 군, 군사법원, 군인, 규명, 규명, 규명, 균형, 근로, 근로자, 금지, 기간제 및 단시간근로자 보호 등에 관한 법률, 기관, 기관, 기관, 기능, 기반, 기법, 기본법, 기본법, 기본법, 기부금, 기술, 기술신용보증기금, 남녀, 노무현, 노인, 노인, 노인, 노인, 농수산물, 농어, 뉴스, 대, 대, 대기, 대책, 대책, 대책, 대책, 대책, 대책, 대책, 대통령, 도로, 도로교통법, 도로교통법, 도법, 도스, 도시, 도시, 도시가스, 도청, 독점규제 및 공정거래에 관한 법률, 독점규제 및 공정거래에 관한 법률, 독점규제 및 공정거래에 관한 법률, 동, 등, 등, 등, 등, 등, 등, 등, 등, 등, 등, 등, 등의, 디, 디자인, 모집, 무역, 문화권, 물질, 민간, 민간인, 민법, 민사소송법, 민사집행법, 박원순, 발전, 발전, 발전, 방송, 방송법, 방송법, 방송법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법안, 법안, 법안, 법안, 병역법, 병원, 병원, 보건, 보상, 보장, 보전, 보전, 보전, 보조, 보험, 보호, 보호, 보호, 보호, 보호, 보호, 복지, 복지, 복지, 복지, 부동산, 부정경쟁방지 및 영업비밀보호에 관한 법률, 분산, 분야, 분야, 분쟁, 불법, 사건, 사건, 사고, 사립대학, 사립학교, 사법, 사업, 사업, 사용, 사이버, 사찰, 사회, 산업, 산업, 산업, 산업, 산업, 산업, 산업, 산지, 상법, 상법, 상생, 상생, 상표, 생산품, 생식, 생태계, 서비스, 서울대학교치과병원, 선거관리위원회, 선거관리위원회, 선거법, 선거법, 선거법, 선거법, 선거법, 선거법, 선거법, 선거법, 설립, 설립, 설립, 설립, 설치, 설치, 설치, 설치, 성폭력범죄의 처벌 등에 관한 특례법, 세무사, 세법, 세법, 소기업, 소기업, 소기업, 소비, 소상, 소상, 소상, 소유, 수, 수당, 수산업협동조합, 수질, 습지, 승강기, 시설, 시설, 식물, 식품, 식품위생법, 신라, 신용보증기금, 신용정보의 이용 및 보호에 관한 법률, 실용신안, 실종, 아동, 아동복지법, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안전, 안전, 안전사고, 야생, 약사법, 양립, 어촌, 어항, 업, 여성, 여신전문금융업법, 여신전문금융업법, 여신전문금융업법, 연구, 연구, 연구, 연안, 영유아보육법, 영유아보육법, 예금, 예방, 예방, 예방, 예방, 예방, 예방, 예방, 예방, 예방, 예방, 예방, 예우, 우선, 우체국, 운영, 운영, 운영, 운영, 원산지, 위생, 위원회, 유교, 유통, 유통, 유통, 유해, 육성, 육성, 육성, 육성, 육성, 의료, 의료, 의료기, 의료법, 의혹, 이명박 정부, 이용, 이전, 이전, 이행, 인터넷, 일, 일, 임대, 임대, 임명, 임명, 임명, 임산부, 입지, 입지, 자립, 자유, 장애인, 장애인, 재난, 재보궐선거, 저작권법, 전기통신사업법, 전원, 전직, 전파, 정보통신망 이용촉진 및 정보보호 등에 관한 법률, 정보통신망 이용촉진 및 정보보호 등에 관한 법률, 정보통신망 이용촉진 및 정보보호 등에 관한 법률, 정부, 정부, 정부, 제한, 제한, 조성, 조세, 조정, 조치, 조치, 조치, 주, 주, 중, 중독, 중등, 중소기업, 중소기업, 중소기업, 중앙, 중앙, 중증, 증진, 지방, 지방교육재정교부금법, 지방세, 지방세법, 지방자치법, 지역, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지위, 지위, 지하수, 진상, 진상, 진상, 진흥, 집적, 집합, 채취, 처벌, 철도, 청소년, 체크카드, 초, 초, 촉진, 촉진, 촉진, 축산물, 출연, 출연, 출연, 택법, 택법, 테러, 통신, 통신, 통신비밀보호법, 퇴거, 투자, 특례, 특례, 특례, 특례법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특정, 특허법, 파견, 편의, 평등, 평생, 폐기물관리법, 폭력, 폭력, 폭력, 폭력, 폭력, 폭력, 폭력, 폭력행위 등 처벌에 관한 법률, 표시, 품, 피해, 하수, 하천, 학교, 학교, 학교, 학교, 학교, 학교, 학교, 학교, 학교, 학생, 학치과, 한국, 한국선박금융, 한국전기안전공사, 한나라당, 항공, 해소, 행정절차법, 향상, 향상, 협력, 협력, 협정, 형법, 홈페이지, 홈페이지, 화학, 환경, 활성,
