--- src/wordcloud_data/billsNouns/19th-2016-nouns.txt
@@ -1,1 +1,0 @@
-가사, 가정, 가족관계의 등록 등에 관한 법률, 가족관계의 등록 등에 관한 법률, 각급, 개량, 개선, 건, 건강, 건널목, 건설, 건축법, 계량, 계획법, 고등, 고용, 공공, 공공, 공공, 공공, 공공, 공공, 공공, 공공, 공무원, 공원, 공직, 공직, 관, 관리, 관리, 관리법, 관할, 교육, 교육, 교육, 교육기본법, 교육법, 교육법, 교육법, 교육법, 교육법, 구역, 국가, 국가, 국가, 국공립, 국공립, 국립대학, 국민건강보험, 국민보호와 공공안전을 위한 테러방지법, 국민연금, 국유재산, 국유재산, 국제, 국회법, 국회법, 국회법, 국회의원, 규명, 규제, 근로기준법, 근로기준법, 근로자, 금지, 기관, 기관, 기금법, 기본, 기본, 기본법, 기본법, 기본법, 기본법, 기업, 기지, 날, 노인, 농, 농, 농업, 농지, 단체, 단체, 단체, 대, 데이트, 도로, 도로교통법, 도시, 도시, 독점규제 및 공정거래에 관한 법률, 동물보호법, 등, 등, 등, 등, 등, 등, 등, 등, 등, 등, 등, 등, 등, 등, 등, 등, 등, 무역, 무역, 물질, 미군, 민간인, 민법, 민사, 민사소송법, 민사조정, 발전, 방송법, 방지, 방지, 범죄, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법안, 법안, 법안, 법원, 법원, 법인, 법인, 벤처기업, 병원, 보건, 보상, 보수, 보장, 보전, 복권, 복권, 복지, 불법, 사건, 사이버, 사회, 사회, 사회복지사, 산업, 산업, 상생, 상속세, 생태계, 서비스, 서울대학교, 선거법, 선거법, 설립, 설립, 설인, 설치, 설치, 세법, 세법, 소송, 수당, 수도권, 심의, 아동복지법, 아동복지법, 아동학대범죄의 처벌 등에 관한 특례법, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 약사법, 어업, 어업, 업, 여신전문금융업법, 연금, 예우, 외식, 운영, 운영, 운영, 운영, 운행, 유공자, 유료, 유아교육법, 육성, 육성, 육성, 육성, 육성, 의과대학, 의과대학, 의료, 의료, 의료, 의료, 의료법, 의료법, 이민, 이전, 이전, 인지, 임산부, 임원, 입양, 자금, 자동차, 자유, 자유, 자치, 장애인, 재생, 재외국민, 재정, 재정, 전담, 전담, 전략산업, 전후, 정비, 정비, 정책, 정치자금법, 제정, 제한, 제한, 제한, 제한, 제한, 제한, 조성, 조세, 조세, 조세, 조세, 조직법, 조치, 존, 주거, 주택, 주택, 주한, 중등, 중등, 중등, 중등, 중소기업, 증여, 증진, 지방, 지방, 지방, 지방교육재정교부금법, 지방교육재정교부금법, 지방교육재정교부금법, 지방세, 지방세, 지방세, 지역, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지정, 진상, 진흥, 진흥, 참전, 처벌, 첨단, 청소년, 체결, 체결, 초, 초, 초, 초, 촉진, 촉진, 태권도, 태권도, 택법, 테러, 통계, 통신비밀보호법, 투자, 특례, 특례, 특례, 특례, 특례, 특례, 특례법, 특례법, 특례법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별회계, 특허법, 편의, 평택시, 폭력, 프리, 한국전쟁, 한국지방재정공제회, 한국청소년연맹, 해양, 해외, 행위, 혁신도시, 협력, 협정, 협정, 형사소송법, 형사소송법, 화재예방, 소방시설 설치·유지 및 안전관리에 관한 법률, 화학, 환경, 희생,
