--- src/wordcloud_data/billsNouns/20th-2020-nouns.txt
@@ -1,1 +1,0 @@
-19, 가정, 가족관계의 등록 등에 관한 법률, 감염병, 감염병, 감염병, 감염병, 감염병, 감염병, 감염병, 감염병, 감염병, 감염병, 감염병, 감염병, 감염병, 감염병, 감염증, 강력범죄, 개선, 거래, 건물, 건설, 건설, 건설, 건설기계, 건축법, 검역, 경상북도, 경제자유구역, 경찰공제회, 고려인, 고용, 공공, 공급, 공동, 공동, 공무원, 공무원법, 공무원법, 공무원법, 공원, 공익사업, 공장, 공정, 공직, 공직, 공직, 공직, 공항, 과학기술, 과학기술인공제회, 관리, 관리, 관리, 관리, 관리, 관리, 관리, 관리, 관리, 관리, 관리, 관리, 관리, 관리, 관리, 관리, 관리, 관리, 관리, 관리, 관리, 관리, 관리, 관리법, 관리법, 관리법, 관리법, 광고, 교류, 교부, 교육기본법, 교육법, 교육법, 구급, 구제, 구조, 국가, 국가, 국가, 국가, 국립대, 국립대학, 국민연금, 국민연금, 국세, 국유재산, 국제연합, 국회, 국회, 국회법, 국회법, 국회법, 국회법, 국회법, 국회의원, 군대, 군사, 군사, 군인, 군인공제회, 권법, 규제, 규제, 근로, 근로기준법, 근로자, 근로자, 금융, 급여, 급여, 기관, 기금, 기금법, 기기, 기본, 기본법, 기본법, 기본법, 기본법, 기본법, 기본법, 기본법, 기본법, 기본법, 기부금, 기부금, 기술, 기지, 기질, 긴급, 낚시, 남녀, 남녀, 남북, 남북, 노인, 노인, 노인, 농어촌, 단체, 대구광역시, 대기, 대책, 대한민국, 대한소방공제회, 대한적십자사, 대한지방행정공제회, 도로교통법, 도로교통법, 도시, 도시, 도시, 도시, 도시, 도시, 도자, 독점규제 및 공정거래에 관한 법률, 동물보호법, 동수, 동포, 등, 등, 등, 등, 등, 등, 등, 등, 등, 등, 등, 등, 등, 등, 등, 등, 등재, 말, 모자, 모집, 모집, 문화, 물, 물, 물가, 미군, 미군, 미래, 민간, 민간, 발전, 방송, 방송법, 방송법, 방지, 범위, 범죄, 범죄, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법률, 법무부, 법안, 법안, 법안, 법원, 법원, 법인, 베트남, 변, 병역법, 병역법, 병역법, 병역법, 병원, 병원, 보건, 보건, 보건, 보건, 보건, 보상, 보상, 보상법, 보장, 보장, 보장, 보장, 보전, 보조, 보존, 보험법, 보호, 보호, 복지, 복지, 부가가치, 부가가치, 북한, 북한이탈주민, 분야, 비디오, 비디오, 비무장지대, 비밀, 사건, 사립학교, 사립학교, 사립학교, 사립학교, 사법경찰관리, 사업, 사업, 사업, 사업, 사업, 사업, 사업, 사업, 사용, 사용, 산업, 산업, 산업, 산업, 산업, 산업, 산업, 산업, 상장, 생명, 생활, 생활, 서울대학교병원, 서울대학교치과병원, 선거법, 선거법, 선거법, 선박, 선출, 설립, 설립, 설립, 설비, 설치, 설치, 설치, 설치, 설치, 성과, 성과, 성폭력범죄의 처벌 등에 관한 특례법, 성폭력범죄의 처벌 등에 관한 특례법, 성폭력범죄의 처벌 등에 관한 특례법, 성폭력범죄의 처벌 등에 관한 특례법, 성폭력범죄의 처벌 등에 관한 특례법, 성폭력범죄의 처벌 등에 관한 특례법, 세계유산, 세계유산, 세대, 세법, 세법, 세법, 세법, 소득세법, 소속, 소유, 소음, 소음, 손해배상, 손해배상, 송, 수당, 수익, 수익, 수행, 스마트, 시기, 시설, 시설, 식품위생법, 신용보증기금, 신용정보의 이용 및 보호에 관한 법률, 실명, 아동·청소년의 성보호에 관한 법률, 아동·청소년의 성보호에 관한 법률, 아동·청소년의 성보호에 관한 법률, 아동·청소년의 성보호에 관한 법률, 아동·청소년의 성보호에 관한 법률, 아동·청소년의 성보호에 관한 법률, 아동복지법, 아동복지법, 아동복지법, 아동학대범죄의 처벌 등에 관한 특례법, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안, 안전, 안전, 안전, 안전, 안전, 안전, 안전, 안전, 안정, 안정, 약사법, 양곡, 양립, 여객, 연구, 연구개발, 영유아보육법, 영화, 영화, 예방, 예방, 예방, 예방, 예방, 예방, 예방, 예방, 예방, 예방, 예방, 예방, 예방, 예방, 예방, 예산, 요양, 용품, 운송, 운송, 운수, 운수, 운수, 운영, 운영, 운영, 원자력, 위원회, 유아교육법, 유지, 육성, 육성, 육성, 육성, 윤리, 은닉, 은닉, 응급, 의료, 의료, 의료법, 인사청문회, 인사청문회, 인터넷 멀티미디어 방송사업법, 일, 임대, 임대, 임산부, 입항, 자격, 자동차, 자동차, 자동차, 자동차, 자동차, 자동차관리법, 자본시장과 금융투자업에 관한 법률, 자살, 자연, 자와, 자원, 자원, 자치, 자치회, 장관, 장기, 장애인, 장애인, 재난, 재난, 재난, 재난, 재생, 재생, 재정, 재정, 재해, 재활용, 재활용, 전, 전기, 전기, 전기용품, 전기통신사업법, 전기통신사업법, 전쟁, 전파, 절약, 절약, 정당법, 정보통신망 이용촉진 및 정보보호 등에 관한 법률, 정보통신망 이용촉진 및 정보보호 등에 관한 법률, 정보통신망 이용촉진 및 정보보호 등에 관한 법률, 정보통신망 이용촉진 및 정보보호 등에 관한 법률, 정보통신망 이용촉진 및 정보보호 등에 관한 법률, 정부, 정착, 정착, 정책, 정치, 제대, 제한, 제한, 제한, 제한, 제한, 제한, 제한, 제한, 제한, 제한, 제한, 조사, 조사, 조성, 조성, 조세, 조세, 조세, 조세, 조세, 조세, 조세, 조세, 조직법, 조직법, 조직법, 존중, 주거, 주민, 주변, 주변, 주택, 주택, 주택, 주택, 주택, 주한, 주한, 중등, 중등, 중립, 중소기업, 증진, 지방, 지방, 지방, 지방, 지방세, 지방세, 지방세, 지방세법, 지방세법, 지역, 지역, 지역, 지역, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지원, 지원금, 지정, 지정, 지하, 직무, 직무, 진흥, 진흥, 진흥, 진흥, 진흥, 진흥, 진흥, 진흥, 집적, 집합, 징수, 참여, 처, 처리, 처벌, 처벌, 처벌, 철도, 철도, 청소년, 청원법, 체류, 체험학습, 초, 초, 촉진, 촉진, 촉진, 촉진, 촉진, 총포·도검·화약류 등의 안전관리에 관한 법률, 출연, 출입국관리법, 출입국관리법, 출입국관리법, 출항, 취득, 취득, 코로나바이러스, 클러스터, 토지, 통신, 통신비밀보호법, 통신비밀보호법, 특례, 특례, 특례, 특례, 특례, 특례, 특례, 특례, 특례, 특례, 특례, 특례법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특별법, 특정, 특정범죄 가중처벌 등에 관한 법률, 편의, 평가, 평등, 평화, 폐기물, 폐기물관리법, 폐기물관리법, 표시, 품, 피해, 피해, 하도급거래 공정화에 관한 법률, 학교, 학교, 학교, 학치과, 한국교직원공제회, 한국대학교육협의회, 한국사학진흥재단, 한국산업은행, 한국인, 한국인, 한국지방재정공제회, 한국철도시설공단, 한반도, 합법, 항공, 항만, 항만, 항만, 해양, 행정심판법, 현장, 협력, 형법, 형법, 형법, 형사소송법, 형사소송법, 화물, 화물, 환경, 활동, 활성, 활성, 활성, 활용, 활용, 회사법,
