+++ src/wordcloud_data/cloud-nouns-21th.json
@@ -1,103 +1,79 @@
 {
+  "21-2023": [
+    ["조세특례제한법", 164],
+    ["기본법", 115],
+    ["공직선거법", 89],
+    ["지방세특례제한법", 83],
+    ["특례법", 71],
+    ["남녀고용평등과", 60],
+    ["양립", 60],
+    ["일·가정", 60],
+    ["국회법", 57],
+    ["설치", 55],
+    ["도로교통법", 53],
+    ["처벌", 52],
+    ["안전관리", 41],
+    ["재난", 40],
+    ["보호", 38],
+    ["지원법", 36],
+    ["소득세법", 35],
+    ["육성", 35],
+    ["이용촉진", 35],
+    ["정보보호", 35],
+    ["정보통신망", 35],
+    ["진흥법", 35],
+    ["아동복지법", 34],
+    ["근로기준법", 31],
+    ["영유아보육법", 31],
+    ["의료법", 31],
+    ["촉진법", 30],
+    ["동물보호법", 29],
+    ["전기통신사업법", 28],
+    ["특별조치법", 28],
+    ["고용보험법", 27],
+    ["보호법", 26],
+    ["국가재정법", 25],
+    ["국민건강보험법", 25],
+    ["노인복지법", 25],
+    ["예방", 25],
+    ["자동차관리법", 25],
+    ["지위", 25]
+  ],
   "21-2022": [
-    ["조세", 151],
-    ["보호", 122],
-    ["공직", 119],
-    ["선거법", 118],
-    ["촉진", 102],
-    ["산업", 94],
-    ["국가", 92],
-    ["안전", 86],
-    ["운영", 84],
-    ["지방세", 62],
-    ["설치", 61],
-    ["교육", 55],
-    ["육성", 54],
-    ["장애인", 54],
-    ["고용", 52],
-    ["사업", 51],
-    ["소득세법", 49],
-    ["보건", 48],
-    ["지역", 47],
-    ["개발", 46],
-    ["세법", 46],
-    ["환경", 46],
-    ["복지", 45],
-    ["지방", 42],
-    ["정비", 41],
-    ["공무원법", 40],
-    ["국회법", 39],
-    ["방지", 39],
-    ["예우", 39],
-    ["이용", 39],
-    ["설립", 38],
-    ["주택", 37],
-    ["유공자", 35],
-    ["도시", 32],
-    ["예방", 31],
-    ["교육법", 30],
-    ["위원회", 30],
-    ["의료", 30],
-    ["발전", 29],
-    ["전기통신", 29],
-    ["중소기업", 29],
-    ["공공", 28],
-    ["재정", 28],
-    ["기관", 27],
-    ["단체", 27],
-    ["보장", 27],
-    ["처벌", 27],
-    ["금융", 26],
-    ["자동차", 26],
-    ["건설", 25],
-    ["조성", 25],
-    ["가정", 24],
-    ["경찰", 24],
-    ["공정", 24],
-    ["교통", 24],
-    ["도로교통법", 24],
-    ["직업", 24],
-    ["거래", 23],
-    ["국민", 23],
-    ["군인", 23],
-    ["남녀", 23],
-    ["문화재", 23],
-    ["보상", 23],
-    ["양립", 23],
-    ["일", 23],
-    ["평등", 23],
-    ["학교", 23],
-    ["주거", 22],
-    ["형사소송법", 22],
-    ["동물보호", 21],
-    ["사법", 21],
-    ["자치", 20],
-    ["공인", 19],
-    ["규명", 19],
-    ["기술", 19],
-    ["성폭력범죄", 19],
-    ["정부조직법", 19],
-    ["체육", 19],
-    ["활동", 19],
-    ["감염병", 18],
-    ["개인정보", 18],
-    ["범죄", 18],
-    ["시설", 18],
-    ["자원", 18],
-    ["재해", 18],
-    ["차별", 18],
-    ["공동", 17],
-    ["근로기준법", 17],
-    ["사립학교", 17],
-    ["소상", 17],
-    ["정보통신망", 17],
-    ["정보보호", 17],
-    ["진상", 17],
-    ["피해", 17],
-    ["형법", 17],
-    ["법인", 16],
-    ["사건", 16],
-    ["소비자", 16]
+    ["조세특례제한법", 197],
+    ["공직선거법", 157],
+    ["기본법", 92],
+    ["지방세특례제한법", 73],
+    ["처벌", 70],
+    ["소득세법", 62],
+    ["국회법", 56],
+    ["설치", 56],
+    ["특례법", 51],
+    ["보호", 95],
+    ["전기통신사업법", 42],
+    ["도로교통법", 40],
+    ["안전관리", 39],
+    ["진흥법", 39],
+    ["촉진법", 37],
+    ["재난", 36],
+    ["예방", 33],
+    ["예우", 31],
+    ["국가재정법", 30],
+    ["남녀고용평등과", 30],
+    ["양립", 30],
+    ["일·가정", 30],
+    ["형사소송법", 30],
+    ["근로기준법", 29],
+    ["공공기관", 28],
+    ["동물보호법", 27],
+    ["육성", 27],
+    ["국민건강보험법", 26],
+    ["지원법", 26],
+    ["성폭력범죄의", 25],
+    ["이용촉진", 25],
+    ["장애인", 25],
+    ["정보보호", 25],
+    ["정보통신망", 25]
   ],
   "21-2021": [
     ["보호", 223],
