--- src/wordcloud_data/getBillsData.js
@@ -1,37 +1,0 @@
-import {appendFileSync} from "fs";
-import axios from "axios";
-
-getBillsData(16);
-getBillsData(17);
-getBillsData(18);
-
-
-
-async function fetchData(age = 21, page = 1) {
-  let data = await axios
-    .get(
-      `https://open.assembly.go.kr/portal/openapi/nzmimeepazxkubdpn?KEY=138ed104529d44a9ab8575f3f5332436&Type=JSON&AGE=${age}&pSize=1000&pIndex=${page}`,
-    )
-    .then((res) => {
-      let data = res.data.nzmimeepazxkubdpn[1].row;
-      return data;
-    })
-    .catch((err) => console.error("호출 fail -", err.name));
-
-  return data;
-}
-
-function getBillsData(age = 21, page = 1) {
-  fetchData(age, page).then((data) => {
-    if (!data) {
-      return;
-    }
-    data.forEach((elem) => {
-      const billName = elem.BILL_NAME;
-      const date = elem.PROPOSE_DT;
-      appendFileSync(`./${age}th-${date.slice(0, 4)}.txt`, billName.replace("일부개정법률안", "") + " ", "utf8");
-    });
-
-    getBillsData(age, page + 1);
-  });
-}
